chore!: rename project to django-helsinki-notification
Alas, the City of Helsinki does not appreciate pigeons.

django-pulu 01/2025 - 02/2025
Gone but not forgotten.
