+++ README.md
@@ -1,2 +1,2 @@
-# django-pulu
+# django-helsinki-notification
 Notifications for City of Helsinki Django apps 🔔
