+++ tests/manage.py
@@ -7,7 +7,7 @@ import sys
 
 def main():
     """Run administrative tasks."""
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "pulu_tests.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.settings")
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
